+++ lib/node/pl-drivers/src/drivers/download_blob.ts
@@ -252,7 +252,7 @@ export class DownloadDriver implements BlobDriver {
   public getComputableContent(
     res: ResourceInfo | PlTreeEntry
   ): ComputableStableDefined<Uint8Array>{
-    return Computable.make((ctx) => 
+    return Computable.make((ctx) =>
       this.getDownloadedBlob(res, ctx), {
         postprocessValue: (v) => v ? this.getContent(v.handle) : undefined
       }