+++ lib/node/pl-middle-layer/src/network_check/network_check.test.ts
@@ -1,10 +1,10 @@
 import { checkNetwork } from './network_check';
+import { testCredentials } from './test_utils';
 
 // it's too long, enable only when you need to see the report.
 test.skip('should return network connectivity report for a platform endpoint set in PL_ADDRESS env', async () => {
-  expect(process.env.PL_ADDRESS).not.toBeUndefined();
+  const { plEndpoint, plUser, plPassword } = testCredentials();
 
-  const plEndpoint = process.env.PL_ADDRESS!;
-  const report = await checkNetwork(plEndpoint);
+  const report = await checkNetwork(plEndpoint, plUser, plPassword);
   console.log(report);
 }, 80000);