+++ etc/blocks/ui-examples/block/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/milaboratories.ui-examples
 
+## 1.0.244
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.7
+
+## 1.0.243
+
+### Patch Changes
+
+- @milaboratories/milaboratories.ui-examples.workflow@1.1.6
+- @milaboratories/milaboratories.ui-examples.ui@1.3.73
+
 ## 1.0.242
 
 ### Patch Changes