+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @milaboratories/pl-drivers
 
+## 1.5.52
+
+### Patch Changes
+
+- @milaboratories/pl-tree@1.6.1
+
+## 1.5.51
+
+### Patch Changes
+
+- 5277219: Deeper network check with templates for downloading and uploading files and running binaries and Python.
+
 ## 1.5.50
 
 ### Patch Changes