+++ sdk/workflow-tengo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @platforma-sdk/workflow-tengo
 
+## 4.2.2
+
+### Patch Changes
+
+- afe3583: - fix for b-query result unmarshaller requiring data to always be available
+  - macro state definitions for awaitState API: BQueryResultSingle, BQueryResultMulti, BQueryResultMultiNoData, PColumnBundle, PColumnBundleWithPartitions
+  - anchor columns are now available by their ids in column bundle
+  - wf.getBlockId() to retrieve string block id
+
+## 4.2.1
+
+### Patch Changes
+
+- 5277219: Deeper network check with templates for downloading and uploading files and running binaries and Python.
+
 ## 4.2.0
 
 ### Minor Changes